pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argon2-quantum"
version = "0.1.0"
description = "Toy quantum stretch KDF with Argon2"
readme = "README.md"
authors = [{name = "Example", email = "example@example.com"}]
requires-python = ">=3.10"
dependencies = [
    "qiskit==1.0.0",
    "argon2-cffi==23.1.0",
    "boto3==1.34.86",
    "redis==5.0.0",
    "amazon-braket-sdk==1.78.3",
]

[project.scripts]
qsargon2 = "qsargon2:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "B"]
